requirements.txt
python-telegram-bot==20.3
httpx==0.24.1
httpcore==0.17.2
